Fix estimate number headers left
